api/requirements.txt
flask-restplus==0.13.0
python-magic==0.4.22
minio==7.0.2